fix code to get definition loading options
