+++ src/javasources/KTool/src/org/kframework/utils/inject/DefinitionLoadingModule.java
@@ -2,15 +2,11 @@
 package org.kframework.utils.inject;
 
 import java.io.File;
-import java.util.Map;
 
-import org.kframework.backend.java.kil.Definition;
 import org.kframework.backend.java.symbolic.JavaSymbolicBackend;
-import org.kframework.kil.Configuration;
 import org.kframework.kil.Term;
 import org.kframework.kil.loader.Context;
 import org.kframework.kompile.KompileOptions;
-import org.kframework.krun.InitialConfigurationProvider;
 import org.kframework.krun.api.KRun;
 import org.kframework.main.GlobalOptions;
 import org.kframework.utils.BinaryLoader;
@@ -21,21 +17,11 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Provider;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
-import com.google.inject.multibindings.MapBinder;
 
 public class DefinitionLoadingModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        MapBinder<KompileOptions.Backend, KRun> mapBinder = MapBinder.newMapBinder(
-                binder(), KompileOptions.Backend.class, KRun.class);
-        for (KompileOptions.Backend enumVal : KompileOptions.Backend.values()) {
-            if (enumVal.generatesDefinition()) {
-                mapBinder.addBinding(enumVal).to(enumVal.krun());
-            }
-        }
-
-        bind(Term.class).toProvider(InitialConfigurationProvider.class);
     }
 
     @Provides @Singleton
@@ -60,25 +46,5 @@ public class DefinitionLoadingModule extends AbstractModule {
     @Provides
     KompileOptions kompileOptions(Context context) {
         return context.kompileOptions;
-    }
-
-    @Provides
-    KRun getKRun(KompileOptions options, Map<KompileOptions.Backend, Provider<KRun>> map) {
-        return map.get(options.backend).get();
-    }
-
-    @Provides @Singleton
-    Configuration configuration(BinaryLoader loader, Context context, Stopwatch sw) {
-        Configuration cfg = loader.loadOrDie(Configuration.class,
-                new File(context.kompiled, "configuration.bin").getAbsolutePath());
-        sw.printIntermediate("Reading configuration from binary");
-        return cfg;
-    }
-
-    @Provides @Singleton
-    Definition javaDefinition(BinaryLoader loader, Context context) {
-        Definition def = loader.loadOrDie(Definition.class,
-                new File(context.kompiled, JavaSymbolicBackend.DEFINITION_FILENAME).toString());
-        return def;
     }
 }
